15-10-25
- Update: Expanded Armor Pack
Attempt to fix a tooltip crash.
Removed camouflage effect from the loner SKAT Knife (???).
Reduced camouflage effect from ISG outfits by a 1/4 (from 20 like a proper guillie to 15). This is so the guillie outfit (or "backpack", if modded) isn't made obsolete.
- Update: MP7 Re-Imagined
Buff to degradation to be on par with other NATO SMGs, on the basis that this mod properly chambers the MP7 in 4.6mmm (unlike vanilla's custom rechamber for 9mm).