23-11-25
Reinstall the patch for the Thompson Modernized with this update if you use it, to fix text issues.

- Update: Weapon stats fix and rebalance patch
JMerc's M1 & M1A1 carbines. Changes similar to those for the SWP equivalents, plus damage nerfed more on par to that of other .357 guns.
JMerc's M1 Garand Tactical: it will use a vanilla NATO silencer.
New ammo module: Pumkin0's Mosin & SKS stripper clips.
Fixed up and compatible with everything. Do note Mags Redux is not prepared to work with clips and strange behaviour will occur.
If present, the Russian Ammo Expansion module will work with it to load the clips with the new ammo.
New weapon mod modules: JMerc's M1 Garand, M1 Garand Tactical, M1 & M1A1 Carbines Pack.
Besides typical fixes and patches, there are patches for Neko's .30-06 Springfield ammo and Armor Modkits addon.
Mags Redux support from M1 Garand Tactical doesn't seem to working on vanilla, so it's split off as a separate patch (and expanded for .30-06 and duplicated for the basic M1 Garand) as a reference for whoever wants to work with it.
- Update: Colt 357. Color correction and file reduction to the icon file, by hoasd1 at the Anomaly Discord server.